fixed issue #1 and unified ESPNow sender on one class
