+++ lib/common/frame.proto
@@ -4,5 +4,5 @@ import "nanopb.proto";
 
 message Frame {
   required uint32 lenght = 1 [(nanopb).int_size = IS_16];
-  repeated uint32 data = 2 [(nanopb).int_size = IS_8, (nanopb).max_size = 1];
+  repeated bytes data = 2 [(nanopb).int_size = IS_8, (nanopb).max_size = 1];
 }
